Split StreamHelpers in two files, remove S template
