+++ cpp/include/Ice/LocalException.h
@@ -5,10 +5,11 @@
 #ifndef ICE_LOCAL_EXCEPTION_H
 #define ICE_LOCAL_EXCEPTION_H
 
+#include "EncodingVersion.h"
 #include "Exception.h"
 #include "Ice/BuiltinSequences.h"
 #include "Ice/Identity.h"
-#include "Protocol.h"
+#include "Ice/Version.h"
 
 #include <string>
 #include <string_view>
