+++ cpp/src/Ice/InstrumentationI.cpp
@@ -11,6 +11,8 @@
 #include <Ice/Communicator.h>
 #include <Ice/LoggerUtil.h>
 
+#include "Ice/Protocol.h"
+
 using namespace std;
 using namespace Ice;
 using namespace IceInternal;
